rg-17: + create useClickOutside hook
